# backend/api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Used only by /api/sonarr/webhook, since Sonarr/Radarr's built-in Webhook connection type has no custom-header support. Any username works; the password must be the API key.",
        "title": "Aura API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/mediaserver/libraries/options": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of library sections from a specified media server configuration. This endpoint accepts a media server configuration in the request body and returns the available library sections for that media server, allowing clients to display options for users to select which library section they want to interact with.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MediaServer"
                ],
                "summary": "Get Library Section Options",
                "parameters": [
                    {
                        "description": "Media Server Configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes_ms.getLibrarySectionOptionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/routes_ms.getLibrarySectionOptionsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized (only when Auth.Enabled=true)",
                        "schema": {
                            "$ref": "#/definitions/httpx.UnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.JSONResponse"
                        }
                    }
                }
            }
        },
        "/api/mediaserver/library/items": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve items from a specific library section in the media server. This endpoint accepts query parameters to identify the library section and pagination options, and returns the items contained within that library section, allowing clients to display the media items available in the selected library section.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MediaServer"
                ],
                "summary": "Get Library Section Items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Library Section ID",
                        "name": "section_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Library Section Title",
                        "name": "section_title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Library Section Type",
                        "name": "section_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start Index for Pagination",
                        "name": "section_start_index",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/routes_ms.GetLibrarySectionItems_Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized (only when Auth.Enabled=true)",
                        "schema": {
                            "$ref": "#/definitions/httpx.UnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "config.Config_MediaServer": {
            "type": "object",
            "properties": {
                "api_token": {
                    "description": "Authentication token for accessing the media server.",
                    "type": "string"
                },
                "enable_plex_event_listener": {
                    "description": "Whether to enable the Plex event listener for reapplying images on refresh. Plex exclusive feature.",
                    "type": "boolean"
                },
                "enable_sort_by_episode_added_date": {
                    "description": "Whether to check episodes for added date when getting Media Items. This is only for Plex servers.",
                    "type": "boolean"
                },
                "libraries": {
                    "description": "List of media server libraries to manage.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LibrarySection"
                    }
                },
                "type": {
                    "description": "Type of media server (e.g., plex, emby, jellyfin).",
                    "type": "string"
                },
                "url": {
                    "description": "Base URL of the media server. This is either the IP:Port or the domain name (e.g., plex.domain.com).",
                    "type": "string"
                },
                "user_id": {
                    "description": "User ID for accessing the media server. This is used for Emby and Jellyfin servers.",
                    "type": "string"
                }
            }
        },
...
        "httpx.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/logging.LogErrorInfo"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "httpx.UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid or expired token"
                }
            }
        },
...
        "logging.LogErrorInfo": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "object"
                },
                "function": {
                    "type": "string",
                    "example": "Reload"
                },
                "help": {
                    "type": "string",
                    "example": "Check server logs for more details"
                },
                "line_number": {
                    "type": "integer",
                    "example": 42
                },
                "message": {
                    "type": "string",
                    "example": "Invalid or expired token"
                }
            }
        },
...
        "models.DBSavedSet": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "selected_types": {
                    "$ref": "#/definitions/models.SelectedTypes"
                },
                "user_created": {
                    "type": "string"
                }
            }
        },
...
        "models.LibrarySection": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Unique identifier for the library section.",
                    "type": "string"
                },
                "media_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MediaItem"
                    }
                },
                "paths": {
                    "description": "Paths of the library section on the media server.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "description": "Title of the library section.",
                    "type": "string"
                },
                "total_size": {
                    "type": "integer"
                },
                "type": {
                    "description": "\"movie\" or \"show\"",
                    "type": "string"
                }
            }
        },
        "models.MediaItem": {
            "type": "object",
            "properties": {
                "added_at": {
                    "description": "Added timestamp",
                    "type": "integer"
                },
                "content_rating": {
                    "description": "Used in MediaItem Details Page - For more information about the item",
                    "type": "string"
                },
                "db_saved_sets": {
                    "description": "Used in MediaItem Details Page - For poster sets",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DBSavedSet"
                    }
                },
                "edition": {
                    "description": "Edition of the media item (e.g., \"Director's Cut\")",
                    "type": "string"
                },
                "guids": {
                    "description": "Used in MediaItem Details Page - For ratings",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MediaItemGuid"
                    }
                },
                "has_mediux_sets": {
                    "description": "Used in Home Page - sorting and filtering",
                    "type": "boolean"
                },
                "ignored_in_db": {
                    "description": "Whether the item is marked as ignored",
                    "type": "boolean"
                },
                "ignored_mode": {
                    "description": "Mode of ignoring (e.g., \"always\", \"until-set-available\", \"until-new-set-available\")",
                    "type": "string"
                },
                "ignored_sets": {
                    "description": "List of set IDs that were present when the item was ignored (used for \"until-new-set-available\" mode)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "latest_episode_added_at": {
                    "type": "integer"
                },
                "library_title": {
                    "description": "Title of the library/section the item belongs to",
                    "type": "string"
                },
                "movie": {
                    "description": "Present if Type is \"movie\"; Contains file info",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.MediaItemMovie"
                        }
                    ]
                },
                "rating_key": {
                    "description": "RatingKey is the internal ID from the media server",
                    "type": "string"
                },
                "released_at": {
                    "description": "Release date timestamp",
                    "type": "integer"
                },
                "series": {
                    "description": "Present if Type is \"show\"; Contains seasons and episodes info",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.MediaItemSeries"
                        }
                    ]
                },
                "summary": {
                    "description": "Summary or description of the media item",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the media item",
                    "type": "string"
                },
                "tmdb_id": {
                    "description": "TMDB ID of the media item",
                    "type": "string"
                },
                "type": {
                    "description": "\"movie\" or \"show\"",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Last updated timestamp",
                    "type": "integer"
                },
                "year": {
                    "description": "Release year of the media item",
                    "type": "integer"
                }
            }
        },
        "models.MediaItemEpisode": {
            "type": "object",
            "properties": {
                "added_at": {
                    "type": "integer"
                },
                "episode_number": {
                    "type": "integer"
                },
                "file": {
                    "$ref": "#/definitions/models.MediaItemFile"
                },
                "rating_key": {
                    "type": "string"
                },
                "season_number": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.MediaItemFile": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "models.MediaItemGuid": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "rating": {
                    "type": "string"
                }
            }
        },
        "models.MediaItemMovie": {
            "type": "object",
            "properties": {
                "file": {
                    "$ref": "#/definitions/models.MediaItemFile"
                }
            }
        },
        "models.MediaItemSeason": {
            "type": "object",
            "properties": {
                "episodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MediaItemEpisode"
                    }
                },
                "rating_key": {
                    "type": "string"
                },
                "season_number": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.MediaItemSeries": {
            "type": "object",
            "properties": {
                "episode_count": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "season_count": {
                    "type": "integer"
                },
                "seasons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MediaItemSeason"
                    }
                }
            }
        },
...
        "models.SelectedTypes": {
            "type": "object",
            "properties": {
                "backdrop": {
                    "type": "boolean"
                },
                "poster": {
                    "type": "boolean"
                },
                "season_poster": {
                    "type": "boolean"
                },
                "special_season_poster": {
                    "type": "boolean"
                },
                "titlecard": {
                    "type": "boolean"
                }
            }
        },
...
        "routes_ms.GetLibrarySectionItems_Response": {
            "type": "object",
            "properties": {
                "library_section": {
                    "$ref": "#/definitions/models.LibrarySection"
                }
            }
        },
...
        "routes_ms.getLibrarySectionOptionsRequest": {
            "type": "object",
            "properties": {
                "media_server": {
                    "$ref": "#/definitions/config.Config_MediaServer"
                }
            }
        },
        "routes_ms.getLibrarySectionOptionsResponse": {
            "type": "object",
            "properties": {
                "library_sections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LibrarySection"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for programmatic/integration access. Generate one under Settings \u003e Authentication, then send it as this header on every request. This is the intended auth method for scripts and integrations - the session cookie above is for browser use only.",
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "SessionCookie": {
            "description": "Browser session cookie, set by POST /api/login or the OIDC callback. Not usable directly via Swagger \"Authorize\" - log in through the app UI in the same browser tab.",
            "type": "apiKey",
            "name": "aura_session",
            "in": "cookie"
        }
    }
}